pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playchitect"
version = "0.1.0"
description = "Smart DJ playlist management with intelligent BPM clustering"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL-3.0"}
authors = [
    {name = "James Westwood"}
]
keywords = ["dj", "playlist", "audio", "clustering", "bpm", "music"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
]

dependencies = [
    "librosa>=0.10.0",
    "mutagen>=1.47.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gui = [
    "PyGObject>=3.42.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.15.0",
    "ty>=0.0.17",
    "pre-commit>=3.3.0",
    "soundfile>=0.12.0",  # For unit tests with synthetic audio
]

[project.scripts]
playchitect = "playchitect.cli.commands:main"
playchitect-gui = "playchitect.gui.app:main"

[project.urls]
Homepage = "https://github.com/james-westwood/playchitect"
Issues = "https://github.com/james-westwood/playchitect/issues"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ty is in beta (0.0.x) — configuration may change between versions.
# Use `uv run ty check` or the pre-commit hook to run type checks.
[tool.ty.environment]
python-version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=playchitect --cov-report=html --cov-report=term"
